Some refactorings

+++ contracts/minter/src/state.rs
@@ -8,7 +8,7 @@ use cw_storage_plus::{Item, Map};
 pub struct Config {
     pub admin: Addr,
     pub base_token_uri: String,
-    pub num_tokens: u64,
+    pub num_tokens: u32,
     pub sg721_code_id: u64,
     pub unit_price: Coin,
     pub whitelist: Option<Addr>,
@@ -18,6 +18,6 @@ pub struct Config {
 
 pub const CONFIG: Item<Config> = Item::new("config");
 pub const SG721_ADDRESS: Item<Addr> = Item::new("sg721_address");
-pub const MINTABLE_TOKEN_IDS: Map<u64, bool> = Map::new("mt");
-pub const MINTABLE_NUM_TOKENS: Item<u64> = Item::new("mintable_num_tokens");
+pub const MINTABLE_TOKEN_IDS: Map<u32, bool> = Map::new("mt");
+pub const MINTABLE_NUM_TOKENS: Item<u32> = Item::new("mintable_num_tokens");
 pub const MINTER_ADDRS: Map<Addr, u32> = Map::new("ma");
